server.ts clean up
